Improve quick list indenting (by matching regex rather than scope)
This one relies on a lovingly hand-crafted regex rather than syntax of
the default Markdown.tmLanguage. Of course this should _really_ be
addressed at the level of the syntax. But I digress:

It supports the following list types (more or less in the order in
which they appear in the regex):

    -
    +
    *
    5.
    a.
    A.
    5)
    a)
    A)
    (5)
    (a)
    (A)
    (@)
    (@label-5_)

Because my attention is starting to wander, it does not support the
[two spaces after a single capital letter followed by a period](http://johnmacfarlane.net/pandoc/README.html#fn1)
rule, as well as roman numerals (e.g., `v)` or `III.`)

There are probably other issues I haven't noticed as well. Such as
those created by this excepionally long commit message :)